% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelids.R
\name{alpaca}
\alias{alpaca}
\title{alpaca species}
\usage{
alpaca(
  name = "Larry",
  strength = 10,
  perception = 10,
  endurance = 10,
  charisma = 10,
  intelligence = 10,
  agility = 10,
  luck = 10,
  level = 1,
  hp = 12,
  species = "alpaca",
  subclass = "ranger",
  special = "pack scout",
  info = "keeps the herd aware of and avoids danger",
  is_human = FALSE,
  model = "smollm2"
)
}
\description{
alpaca species
}
