% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelids.R
\name{camelids}
\alias{camelids}
\title{a base class for camelids}
\usage{
camelids(
  name = "Larry",
  strength = 10,
  perception = 10,
  endurance = 10,
  charisma = 10,
  intelligence = 10,
  agility = 10,
  luck = 10,
  level = 1,
  hp = 12
)
}
\description{
a base class for camelids
}
