% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crias-package.R
\docType{package}
\name{crias-package}
\alias{crias}
\alias{crias-package}
\title{crias: a DnD game in R made possible by ollama and ellmer}
\description{
The package is designed as a turn-based game between human and AI players. But instead of Dungeons and Dragons, it is Crags y Crias.
}
\author{
\strong{Maintainer}: Frank Hull \email{frankiethull@gmail.com}

}
\keyword{internal}
