% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\docType{data}
\name{d10}
\alias{d10}
\title{a 10-sided die tool}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
d10
}
\description{
a 10-sided die tool
}
\keyword{datasets}
