% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\docType{data}
\name{d12}
\alias{d12}
\title{a 12-sided die tool}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
d12
}
\description{
a 12-sided die tool
}
\keyword{datasets}
