% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\docType{data}
\name{d20}
\alias{d20}
\title{a 20-sided die tool}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
d20
}
\description{
a 20-sided die tool
}
\keyword{datasets}
