% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\docType{data}
\name{d6}
\alias{d6}
\title{a 6-sided die tool}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
d6
}
\description{
a 6-sided die tool
}
\keyword{datasets}
