% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\docType{data}
\name{d8}
\alias{d8}
\title{a 8-sided die tool}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
d8
}
\description{
a 8-sided die tool
}
\keyword{datasets}
