% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelids.R
\name{guanaco}
\alias{guanaco}
\title{guanaco species}
\usage{
guanaco(
  name = "Larry",
  strength = 10,
  perception = 10,
  endurance = 10,
  charisma = 10,
  intelligence = 10,
  agility = 10,
  luck = 10,
  level = 1,
  hp = 12,
  species = "guanaco",
  subclass = "survivalist",
  special = "climbing and trekking",
  info = "high probability of safe routes and survival",
  is_human = FALSE,
  model = "tinyllama"
)
}
\description{
guanaco species
}
