% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelids.R
\name{llama}
\alias{llama}
\title{llama species}
\usage{
llama(
  name = "Larry",
  strength = 10,
  perception = 10,
  endurance = 10,
  charisma = 10,
  intelligence = 10,
  agility = 10,
  luck = 10,
  level = 1,
  hp = 12,
  species = "llama",
  subclass = "guardian",
  special = "pack healer",
  info = "keeps the herd together and safe",
  is_human = TRUE,
  model = "none"
)
}
\description{
llama species
}
