% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pachamama.R
\name{pachamama}
\alias{pachamama}
\title{pachamama, the story teller}
\usage{
pachamama(
  model = "mistral",
  system_prompt =
    "you are a terse dungeon master named Pachamama. \\n\\n                                                              you are designed to moderate and create gameplay. \\n\\n                                                              you guide a herd of mythical young llamas through the andes. \\n\\n                                                              this is a fantasy andean world, \\n\\n                                                              you may introduce monsters or themes based on the inca, andes, and amazon. \\n\\n                                                              you determine the setting, speed, and all aspects \\n\\n                                                              of this llama-centric adventure story. \\n\\n\\n                                                              speak english unless prompted otherwise. \\n\\n                                                              you will create a unique quest for each initialized a turn-based game. \\n\\n                                                              players will be required to roll dice once the game starts. \\n\\n                                                              you will decide when and which dice to use.\\n                                                             ",
  initial_prompt =
    "hello pachamama, \\n\\n                                                              please provide a backstory and present a quest (<100 tokens). \\n\\n                                                              I'll either accept or deny and register my party.\\n\\n                                                              THEN once we have started a game, we can roll dice.",
  tools = list(<object>, <object>, <object>, <object>, <object>, <object>)
)
}
\description{
pachamama, the story teller
}
