% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_game}
\alias{play_game}
\title{Play Game}
\usage{
play_game(
  pachamama = crias::pachamama(),
  herd = crias::herd(characters = list(llama = crias::llama(name = "lance"), vicuna =
    crias::vicuna(name = "victor"))),
  gamelog = character(0),
  rounds = 3
)
}
\arguments{
\item{pachamama}{a pachamama}

\item{herd}{a herd of characters}

\item{gamelog}{a gamelog to continue where the round left off.}

\item{rounds}{num of rounds}
}
\value{
a gamelog
}
\description{
Play Game
}
