% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_round}
\alias{play_round}
\title{Play a Round}
\usage{
play_round(
  pachamama = crias::pachamama(),
  herd = crias::herd(characters = list(llama = crias::llama(name = "lance"), vicuna =
    crias::vicuna(name = "victor"))),
  gamelog = character(0)
)
}
\arguments{
\item{pachamama}{a pachamama}

\item{herd}{a herd of characters}

\item{gamelog}{a gamelog to continue where the round left off.}
}
\value{
a gamelog
}
\description{
Play a Round
}
