% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelids.R
\name{vicuna}
\alias{vicuna}
\title{vicuna species}
\usage{
vicuna(
  name = "Larry",
  strength = 10,
  perception = 10,
  endurance = 10,
  charisma = 10,
  intelligence = 10,
  agility = 10,
  luck = 10,
  level = 1,
  hp = 12,
  species = "vicuna",
  subclass = "rogue",
  special = "stealth and camouflage",
  info = "quick, quiet, and evades easily",
  is_human = FALSE,
  model = "tinyllama"
)
}
\description{
vicuna species
}
